Updated kwarg behavior of functions in Fit.
Functions in Fit like constrain_parameter_pair() now take params as a keyword argument. If no value is given (params=None), then self.params is used as the Parameters instance to modify. This behavior should allow the methods in Fit to be conveniently applied to a Parameters instance within a Fit instance and will also enable the methods to be used more generically on semi-arbitrary Parameters instances that do not necessarily belong to a Fit instance.